Rename BuildablePropertyFactory.CodeGenerator -> BuildableProperty
BuildablePropertyFactory is now BuildableProperty.Factory
